Fixed Duplicate User Bug
The file and string needs to be inside the function as it needs to be
checked each time the fucntion is run and not when the server first
starts. This allows it to constantly update the registerusers array
every time a new user is called.  This came to me in a dream and we
fixed it in the dream...